requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
pymysql>=1.1.0
sqlalchemy>=2.0.25
alembic>=1.13.0

# Cryptography
cryptography>=42.0.0
pycryptodome>=3.23.0

# Cache
redis>=5.0.0

# Environment Variables
python-dotenv>=1.0.0

# OAuth
authlib>=1.6.6
httpx>=0.26.0

# Security
itsdangerous>=2.1.2
PyJWT>=2.8.0

# AI/ML
openai
# sentence-transformers>=2.2.0
# qdrant-client>=1.7.0

# Settings & Validation
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
pydantic-settings>=2.12.0
email_validator

# AWS & Storage
boto3>=1.34.0
Pillow>=10.2.0
python-multipart>=0.0.21